+++ lib/src/main/java/net/ypresto/androidtranscoder/engine/MediaTranscoderEngine.java
@@ -24,9 +24,11 @@ import android.os.Build;
 import net.ypresto.androidtranscoder.MediaTranscoderOptions;
 import net.ypresto.androidtranscoder.source.DataSource;
 import net.ypresto.androidtranscoder.strategy.OutputStrategyException;
+import net.ypresto.androidtranscoder.transcode.AudioTrackTranscoder;
 import net.ypresto.androidtranscoder.transcode.NoOpTrackTranscoder;
 import net.ypresto.androidtranscoder.transcode.PassThroughTrackTranscoder;
 import net.ypresto.androidtranscoder.transcode.TrackTranscoder;
+import net.ypresto.androidtranscoder.transcode.VideoTrackTranscoder;
 import net.ypresto.androidtranscoder.utils.ISO6709LocationParser;
 import net.ypresto.androidtranscoder.utils.Logger;
 import net.ypresto.androidtranscoder.validator.Validator;
