+++ lib/src/main/java/net/ypresto/androidtranscoder/remix/PassThroughAudioRemixer.java
@@ -1,0 +1,15 @@
+package net.ypresto.androidtranscoder.remix;
+
+import java.nio.ShortBuffer;
+
+/**
+ * The simplest {@link AudioRemixer} that does nothing.
+ */
+public class PassThroughAudioRemixer implements AudioRemixer {
+
+    @Override
+    public void remix(final ShortBuffer inSBuff, final ShortBuffer outSBuff) {
+        // Passthrough
+        outSBuff.put(inSBuff);
+    }
+}
